Implemented Searchbar inside SongList component
