+++ src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import Header from './components/Header';
 import AddSong from './components/AddSong';
-import SearchBar from './components/SearchBar';
 import SongList from './components/SongList';
 import SongPlayer from './components/SongPlayer';
 import { Grid, useMediaQuery } from '@material-ui/core';
@@ -33,7 +32,6 @@ function App() {
         paddingTop: greaterThanSm ? 80 : 20 
        }} item xs={12} md={7}>
       <AddSong />
-      <SearchBar />
       <SongList />
       </Grid>
       <Grid style={
