+++ src/components/SongList.js
@@ -1,16 +1,48 @@
 import React from 'react'
-import { Card, CircularProgress, CardMedia, CardContent, Typography, CardActions, IconButton, makeStyles} from '@material-ui/core';
-import { PlayArrow, Pause } from '@material-ui/icons';
+import { Card, CircularProgress, CardMedia, CardContent, Typography, CardActions, IconButton, makeStyles, TextField, InputAdornment} from '@material-ui/core';
+import { PlayArrow, Pause, Search } from '@material-ui/icons';
 import { useSubscription, useMutation } from '@apollo/react-hooks';
 import { GET_SONGS } from '../graphql/subscriptions';
 import { SongContext } from '../App';
 import { REMOVE_OR_ADD_FROM_PLAYLIST } from '../graphql/mutations';
 import PlaylistAddOutlinedIcon from '@material-ui/icons/PlaylistAddOutlined';
 
+const useStyles = makeStyles( theme => ( {
+    containerSearch: {
+        display: 'flex',
+        alignItems: 'center'
+    },
+    textInput: {
+        margin: theme.spacing(1)
+    },
+    containerSongList: {
+        margin: theme.spacing(3)
+    },
+    songInfoContainer: {
+        display: 'flex',
+        alignItems: 'center'
+    },
+    songInfo: {
+        width: '100%',
+        display: 'flex',
+        justifyContent: 'space-between'
+    },
+    thumbnail: {
+        objectFit: 'cover',
+        width: 140,
+        height: 140
+    }
+}))
+
 function SongList() {
 
     //But now we are subscribing to new data changes
     const { data, loading, error } = useSubscription(GET_SONGS)
+    const classes = useStyles()
+
+    // Integrating Search bar inside SongList component
+    // for easier functionality / Implementation
+    const [searchTerm, setSearchTerm] = React.useState('')
 
     // const song = {
     //     title: 'Abhishek',
@@ -38,35 +70,44 @@ function SongList() {
             </div>
         )
     }
-    
-    return <div>
-        {/* Return each song 10 times */}
-        {/* {Array.from({ length: 10 }, () => song).map((song, i) => ( */}
-        {data.songs.map((song) => (
-            <Song key={song.id} song={song} />
-        ))}
-    </div>
-}
 
-const useStyles = makeStyles( theme => ({
-    container: {
-        margin: theme.spacing(3)
-    },
-    songInfoContainer: {
-        display: 'flex',
-        alignItems: 'center'
-    },
-    songInfo: {
-        width: '100%',
-        display: 'flex',
-        justifyContent: 'space-between'
-    },
-    thumbnail: {
-        objectFit: 'cover',
-        width: 140,
-        height: 140
+    function handleEditSearchTerm(sch) {
+        setSearchTerm(sch.target.value)
     }
-}))
+
+    function handleDynamicSearch() {
+        return data.songs.filter(song => song.title.toLowerCase().includes(searchTerm.toLowerCase()))
+    }
+    
+    return (
+        <React.Fragment>
+        <div className={classes.containerSearch}>
+            {/* <input type='text' placeholder='Search for a song by name!' /> */}
+            <TextField 
+            className = { classes.textInput }
+            value = { searchTerm }
+            onChange = { handleEditSearchTerm }
+            placeholder = "Search for a song by name!" 
+            fullWidth
+            margin="normal" 
+            type="text" 
+            InputProps={{
+                startAdornment: (
+                    <InputAdornment position="start">
+                    <Search/>
+                    </InputAdornment>
+                )
+            }}/>
+        </div>
+        <div>
+            {
+            handleDynamicSearch().map((song) => (
+                <Song key={song.id} song={song} />
+            ))}
+        </div>
+        </React.Fragment>
+    )
+}
 
 function Song({ song }) {
 
@@ -98,7 +139,7 @@ function Song({ song }) {
     }
 
     return (
-        <Card className={classes.container}>
+        <Card className={classes.containerSongList}>
             <div className={classes.songInfoContainer}>
                 <CardMedia className={classes.thumbnail} image={thumbnail} />
                 <div className={classes.songInfo}>
